added examples, changed extraction_CPA fundamentally!, global plot function
extraction_CPA was using the wrong fluence to calculate the saturation, as it only took the saturation of thin disc of a single wavelength, however, we must use the total fluence to calculate the saturation accurately.
The plotting of the data was completely rewritten, there is now a generalized plot function in utilities, which is accessed by every script.